requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyascii==0.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
Babel==2.12.1
brotlipy==0.7.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646038595831/work
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
contourpy==1.0.7
coqpit==0.0.17
cryptography @ file:///C:/b/abs_36x9ifdcl4/croot/cryptography_1665612655344/work
cycler==0.11.0
Cython==0.29.28
datasets==2.8.0
dateparser==1.1.7
decorator==5.1.1
dill==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
evaluate==0.4.0
ffmpeg-python==0.2.0
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
Flask==2.2.3
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.3
g2pkk==0.1.2
gitdb==4.0.10
GitPython==3.1.30
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
huggingface-hub==0.11.1
huggingsound==0.1.6
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562621412/work
importlib-resources==5.12.0
inflect==5.6.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
jiwer==2.5.1
joblib @ file:///C:/b/abs_e60_bwl1v6/croot/joblib_1666298845728/work
jsonlines==1.2.0
kiwisolver==1.4.4
Levenshtein==0.20.2
librosa==0.8.0
llvmlite==0.38.1
loguru==0.5.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mecab-python3==1.0.5
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
more-itertools==9.1.0
MouseInfo==0.1.3
multidict==6.0.2
multiprocess==0.70.14
networkx==2.8.8
nltk==3.8.1
notify-py==0.3.3
num2words==0.5.12
numba==0.55.1
numpy==1.21.6
openai-whisper @ git+https://github.com/openai/whisper.git@ad3250a846fe7553a25064a2dc593e492dadf040
packaging @ file:///C:/b/abs_cfsup8ur87/croot/packaging_1671697442297/work
pandas==1.5.2
pathtools==0.1.2
Pillow==9.2.0
pipreqs==0.4.11
pooch==1.6.0
protobuf==3.19.6
psutil==5.9.4
pyarrow==10.0.1
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynndescent==0.5.8
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyperclip==1.8.2
pypinyin==0.48.0
PyRect==0.2.0
pysbd==0.3.4
PyScreeze==0.1.28
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rapidfuzz==2.13.7
regex @ file:///C:/ci/regex_1658258307256/work
requests @ file:///C:/ci/requests_1657735342357/work
resampy==0.4.2
responses==0.18.0
sacremoses==0.0.43
scikit-learn==1.2.0
scipy==1.9.3
sentry-sdk==1.13.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sounddevice==0.4.6
soundfile==0.11.0
srt==3.5.2
tensorboardX==2.6
threadpoolctl==3.1.0
tokenizers==0.13.0.dev0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.13.0+cu116
torchaudio==0.13.0
torchtext==0.14.0
torchvision==0.14.0
tornado @ file:///C:/ci/tornado_1606924294691/work
tqdm @ file:///C:/b/abs_0axbz66qik/croots/recipe/tqdm_1664392691071/work
trainer==0.0.20
transformers==4.25.1
TTS==0.11.1
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3 @ file:///C:/b/abs_a8_3vfznn_/croot/urllib3_1666298943664/work
vosk==0.3.44
wandb==0.13.9
webrtcvad==2.0.10
websockets==10.4
Werkzeug==2.2.3
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
win32-setctime==1.1.0
wincertstore==0.2
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.1
zipp @ file:///C:/ci/zipp_1652273994994/work
